# Speedup wide integers