… tests/queries/0_stateless/01440_big_int_exotic_casts.reference
… tests/queries/0_stateless/01440_big_int_exotic_casts.sql