I hereby agree to the terms of the CLA available at: https://yandex.ru/legal/cla/?lang=en

Changelog category (leave one):
- Performance Improvement

Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Improve performance of 256-bit bytes using (u)int64_t as base type for wide integers. Original wide integers use 8-bit types as base.

There're also some minor optimizations.

I've tested simple queries. Wide 256-bit integer is about 20 times slower than __int128 with PR applied.
```
SELECT toInt128(number) * number
FROM numbers_mt(10000000000)
FORMAT Null

0 rows in set. Elapsed: 3.709 sec. Processed 10.00 billion rows, 80.00 GB (2.70 billion rows/s., 21.57 GB/s.) 
```
```
SELECT toInt256(number) * number
FROM numbers_mt(10000000000)
FORMAT Null

0 rows in set. Elapsed: 63.936 sec. Processed 10.00 billion rows, 80.00 GB (156.41 million rows/s., 1.25 GB/s.) 
```