@4ertus2 2020-09-17
PVS error looks like false positive: if (bit_shift) => (overflow_shift < 64).