@4ertus2 2020-09-18
Fuzzer error is in fetch phase: wget ERROR 404: Not Found.